Move ShowPdf to __init__.py and add __main__.py
The package contained a single meaningful file, tkPDFViewer.py which
consisted of the class ShowPdf and a main() function to execute an example.
To avoid having to import tkPDFViewer.tkPDFViewer.ShowPdf each time, it makes
more sense to move ShowPdf and main() to  __init__.py while the `if __name__ == "__main__"`
block should be in the __main__.py. Changed main() so that the filepath is not hard coded
but is given in argument through the command line: `python -m tkPDFViewer <pdf file path>`